refactor(option-amm): break down table components
