+++ apps/dapp/src/components/option-amm/Tables/StrikesChain/TableDisclosure.tsx
@@ -1,0 +1,36 @@
+import { Disclosure } from '@dopex-io/ui';
+
+import formatAmount from 'utils/general/formatAmount';
+
+export interface DisclosureStrikeItem {
+  iv: number;
+  delta: number;
+  theta: number;
+  vega: number;
+  gamma: number;
+}
+
+const StatItem = ({ name, value }: { name: string; value: string }) => (
+  <div className="flex flex-col px-1">
+    <span className="text-sm font-medium">{value}</span>
+    <span className="text-stieglitz text-xs">{name}</span>
+  </div>
+);
+
+const TableDisclosure = (props: DisclosureStrikeItem) => {
+  return (
+    <Disclosure.Panel as="tr" className="bg-umbra">
+      <td colSpan={5}>
+        <div className="grid grid-cols-5 gap-6 p-3 border-b border-dashed border-carbon">
+          <StatItem name="IV" value={String(props.iv)} />
+          <StatItem name="Delta" value={formatAmount(props.delta, 5)} />
+          <StatItem name="Vega" value={formatAmount(props.vega, 5)} />
+          <StatItem name="Gamma" value={formatAmount(props.gamma, 5)} />
+          <StatItem name="Theta" value={formatAmount(props.theta, 5)} />
+        </div>
+      </td>
+    </Disclosure.Panel>
+  );
+};
+
+export default TableDisclosure;
